update to uwsgi and settings

+++ conf/install.bash
@@ -26,26 +26,24 @@ sudo pip install Django==1.10.8
 # download greenery app
 sudo git clone https://github.com/jeffleary00/greenery.git /var/www/greenery
 
-# build rfoutlet binary
-# git clone https://github.com/toofishes/rfoutlet-pi.git /tmp/rfoutlet-pi
-# mv /tmp/rfoutlet-pi/Makefile /tmp/rfoutlet-pi/Makefile.bak
-# cp /var/www/greenery/conf/rfoutlet.Makefile /tmp/rfoutlet-pi/Makefile
-# cd /tmp/rfoutlet-pi
-# sudo make
-# sudo -u www-data cp /tmp/rfoutlet-pi/rfoutlet /var/www/greenery/bin
 
 # configure web app
 sudo chown -R www-data /var/www 
 sudo chgrp -R www-data /var/www
-sudo mv /etc/nginx/sites-available/default /etc/nginx/sites-available/default.bak
+if [ -f /etc/nginx/sites-available/default ]; then
+   sudo mv /etc/nginx/sites-available/default /etc/nginx/sites-available/default.bak
+fi
 sudo cp /var/www/greenery/conf/www/nginx.default /etc/nginx/sites-available/default
+if [ -f /etc/nginx/sites-enabled/default ]; then
+   sudo rm /etc/nginx/sites-enabled/default
+fi
 sudo ln -s /etc/nginx/sites-available/default /etc/nginx/sites-enabled/default
 sudo cp /var/www/greenery/conf/www/uwsgi.ini /etc/uwsgi/apps-available/uwsgi.ini
+if [ -f /etc/uwsgi/apps-enabled/uwsgi.ini ]; then
+   sudo rm /etc/uwsgi/apps-enabled/uwsgi.ini
+fi
 sudo ln -s /etc/uwsgi/apps-available/uwsgi.ini /etc/uwsgi/apps-enabled/uwsgi.ini
 
-# inital app db
-# cd /var/www/greenery
-# sudo -u www-data python db_create.py
 
 # restart services
 sudo service uwsgi restart
